feature: AC-301 Dockerfiles and Docker Compose for AdminApp-v4 front-end and backend (#71)
* feat: Add Docker configuration for AdminApp including API and frontend services

* feature: Update Docker configuration and routing for AdminApp with Nginx proxy

* feature: Add local.js-edfi-docker to .gitignore exceptions

* feature: Add local.js-edfi-docker configuration for Nginx proxy

* feature: Enhance OIDC integration and update environment configurations

* feature: Add environment configuration for AdminApp and update Docker setup

* feature: Update AdminApp configuration and scripts for improved OIDC integration and Docker setup

* fix: correct redirect URI in Keycloak configuration and update typo in documentation

* docs: update README to clarify Admin App service start/stop instructions and enhance Dockerfile description

* fix: update Dockerfile.api for improved build process and health check configuration

* fix: remove KEYCLOAK_INTERNAL_URL from environment variables and update VITE_OIDC_ID default value in docker-compose

+++ compose/down.ps1
@@ -9,25 +9,51 @@ param(
 
     # Drop Keycloak's volume as well
     [Switch]
-    $Keycloak
+    $Keycloak,
+
+    # Only Admin App's services
+    [Switch]
+    $AdminApp
 )
 
-if ($V) {
-    # Drop existing volumes
-    docker compose down --volumes
+try {
+    if ($AdminApp) {
+        Push-Location adminapp
+        .\down.ps1
+        Pop-Location
+    }
+    else {
+        if ($V) {
+            # Drop existing volumes
+            docker compose down --volumes
+        }
+        else {
+            # Shuts down the Docker Compose services without dropping volumes
+            docker compose down
+        }
+
+        if ($Keycloak) {
+            if ($V) {
+                # Shut down Keycloak and remove its volume
+                docker compose --project-name edfiadminapp-dev-environment -f keycloak.yml down --volumes
+            }
+            else {
+                # Shut down Keycloak without removing its volume
+                docker compose --project-name edfiadminapp-dev-environment -f keycloak.yml down
+            }
+        }
+        
+        Push-Location adminapp
+        .\down.ps1
+        Pop-Location
+    }
+    Write-Host "SUCCESS! Services stopped successfully." -ForegroundColor Green
+    exit 0
 }
-else {
-    # Shuts down the Docker Compose services without dropping volumes
-    docker compose down
+catch {
+    Write-Host "ERROR! Services failed to stop." -ForegroundColor Red
+    exit 1
 }
 
-if ($Keycloak) {
-  if ($V) {
-    # Shut down Keycloak and remove its volume
-    docker compose --project-name edfiadminapp-dev-environment -f keycloak.yml down --volumes
-  }
-  else {
-      # Shut down Keycloak without removing its volume
-      docker compose --project-name edfiadminapp-dev-environment -f keycloak.yml down
-  }
-}
+
+
